Refactored identification module
