+++ python/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "desyder"
-version = "0.1.0"
+version = "0.1.3"
 description = "ForSyDe's Design Space Exploration Tool"
 authors = ["jordao <jordao@kth.se>"]
 license = "MIT"
